Offer tutorial and usage examples in readme
I've written a long-form guide to using the `routing-duplex` library to serve as a README. It starts with a short summary and usage example, but then walks through several more examples in the style of `heterogeneous`.

There may be some typos or things you'd like to correct about how I've described the library and its use, and I welcome those! I'd love to have this up for others to be able to begin using the library.